+++ src/tool/xmodeler/ControlCenter.java
@@ -59,7 +59,7 @@ import tool.helper.auxilaryFX.JavaFxButtonAuxilary;
 import tool.helper.persistence.StartupModelLoader;
 import tool.helper.user_properties.PropertyManager;
 import tool.helper.user_properties.UserProperty;
-import tool.xmodeler.tool_introduction.ToolIntroductionManager;
+import tool.xmodeler.didactic_ml.ToolIntroductionManager;
 
 public class ControlCenter extends Stage {
